updated brunel network to work with both Python 2 and 3

+++ tests/models/brunel_network_function.py
@@ -33,8 +33,8 @@ def brunel_network(eta, g, delay, J_E):
     N_E = 1000             # Number of inhibitory neurons
     N_I = 250              # Number of excitatory neurons
     N_neurons = N_E + N_I  # Number of neurons in total
-    C_E = N_E/10           # Number of excitatory synapses per neuron
-    C_I = N_I/10           # Number of inhibitory synapses per neuron
+    C_E = int(N_E/10)      # Number of excitatory synapses per neuron
+    C_I = int(N_I/10)      # Number of inhibitory synapses per neuron
     J_I = -g*J_E           # Amplitude of inhibitory postsynaptic current
 
     nu_ex = eta*V_th/(J_E*C_E*tau_m)
@@ -71,6 +71,7 @@ def brunel_network(eta, g, delay, J_E):
     # Connect neurons to each other
     nest.CopyModel('static_synapse_hom_w', 'excitatory',
                    {'weight':J_E, 'delay':delay})
+
     nest.Connect(nodes_E, nodes,
                  {'rule': 'fixed_indegree', 'indegree': C_E},
                  'excitatory')
